hfst-twolc unit tests added.
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@1366 941e2c2b-deac-454f-805a-451daa25f33c

+++ tools/src/hfst-twolc/src/io_src/InputReader.cc
@@ -69,7 +69,7 @@ int main(void)
   std::string str1
     ("Some text spanning one line only.");
   std::istringstream in1(str1);
-  int counter = 1;
+  size_t counter = 1;
   InputReader ir1(counter);
   try
     {
